refactor(ui): address seventh self-review round — consolidate duplicate helpers
- Move deriveInitials to @/lib/format (canonical home); re-export from
  table-editor-view for backwards compat; import from lib in task-grouped-view
- Remove toHumanLabel from process-run-section (same as formatStepName); import
  formatStepName from @/lib/format instead
- Remove resolveStepLabel from process-detail (same as formatStepName); inline
  the two call sites; import formatStepName from @/lib/format instead
- Switch process-detail formatStepName import from task-utils to @/lib/format
