+++ packages/platform-ui/src/lib/format.ts
@@ -24,6 +24,13 @@ export function formatStepName(stepId: string): string {
     .replace(/\b\w/g, (c) => c.toUpperCase());
 }
 
+export function deriveInitials(name: string): string {
+  const words = name.trim().split(/\s+/);
+  const first = words[0]?.[0] ?? '';
+  const last = words.length > 1 ? words[words.length - 1]![0] ?? '' : '';
+  return (first + last).toUpperCase();
+}
+
 export function computeWallClockDuration(
   createdAt: string,
   stepExecutions: StepExecution[],
